+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-20T11:17:25.021Z",
+    "2026-04-20T12:20:36.270Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {