+++ packages/orq-rc/RELEASES.md
@@ -8369,3 +8369,13 @@ Based on:
 - [typescript v4.8.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.29 - packages/orq-rc
+
+## 2026-04-20 12:23:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.8 (2.881.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.30 - packages/orq-rc